refactor games
