+++ lib/router/route_map.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/widgets.dart';
-import 'package:memory_lamp/screens/fill_in_the_blanks.dart';
+import 'package:memory_lamp/screens/games/fill_in_the_blanks.dart';
+import 'package:memory_lamp/screens/games/pick_game.dart';
+import 'package:memory_lamp/screens/games/rearrange_to_memorize.dart';
+import 'package:memory_lamp/screens/games/recite_this_verse.dart';
 import 'package:memory_lamp/screens/home.dart';
 import 'package:memory_lamp/screens/onboarding.dart';
-import 'package:memory_lamp/screens/pick_game.dart';
-import 'package:memory_lamp/screens/rearrange_to_memorize.dart';
-import 'package:memory_lamp/screens/recite_this_verse.dart';
 import 'package:memory_lamp/screens/signup.dart';
 import 'package:memory_lamp/screens/emotions.dart';
