--- lib/screens/fill_in_the_blanks.dart
@@ -1,65 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:memory_lamp/helpers/size_mq.dart';
-import 'package:memory_lamp/models/verse.dart';
-import 'package:memory_lamp/theming/defaults.dart';
-import 'package:memory_lamp/theming/ml_colors.dart';
-import 'package:memory_lamp/theming/ml_font.dart';
-import 'package:memory_lamp/widgets/buttons/ml_text_button.dart';
-import 'package:memory_lamp/widgets/ml_container.dart';
-import 'package:memory_lamp/widgets/ml_text.dart';
-
-class FillInTheBlanksScreen extends StatelessWidget {
-  static const String routeName = "/fillInTheBlanks";
-
-  final Verse _verse;
-
-  const FillInTheBlanksScreen(this._verse, {Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: MLText("Memorize"), centerTitle: true),
-      body: SafeArea(
-        child: Padding(
-          padding: MLDefaults.screenPadding,
-          child: Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                MLText(
-                  _verse.book,
-                  style: MLFont.memorizeHeader,
-                ),
-                MLText(
-                  _verse.verse,
-                  height: 1.8,
-                ),
-                MLTextButton(
-                  width: SizeMQ.width! * .7,
-                  backgroundColor: MLColors.secondary,
-                  child: MLText("Next"),
-                ),
-                MLContainer(
-                  backgroundColor: MLColors.primary,
-                  padding: const EdgeInsets.all(1),
-                  borderRadius: MLDefaults.rounded,
-                  child: ClipRRect(
-                    borderRadius: MLDefaults.rounded,
-                    child: LinearProgressIndicator(
-                      backgroundColor: Colors.white,
-                      valueColor: AlwaysStoppedAnimation<Color>(
-                        MLColors.primary,
-                      ),
-                      value: .20,
-                      minHeight: 12,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
