+++ lib/screens/games/rearrange_to_memorize.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:memory_lamp/theming/defaults.dart';
+import 'package:memory_lamp/widgets/compound_widgets/memorize_scaffold.dart';
 import 'package:memory_lamp/widgets/ml_text.dart';
 
 class RearrangeToMemorizeScreen extends StatelessWidget {
@@ -9,14 +9,8 @@ class RearrangeToMemorizeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: MLText("Memorize"), centerTitle: true),
-      body: SafeArea(
-        child: Padding(
-          padding: MLDefaults.screenPadding,
-          child: MLText("Rearrange to memorize"),
-        ),
-      ),
+    return MemorizeScaffold(
+      body: MLText("Rearrange to memorize"),
     );
   }
 }
