+++ lib/screens/games/recite_this_verse.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:memory_lamp/widgets/compound_widgets/memorize_scaffold.dart';
+import 'package:memory_lamp/widgets/ml_text.dart';
+
+class ReciteThisVerseScreen extends StatelessWidget {
+  static const String routeName = "/reciteThisVerse";
+
+  const ReciteThisVerseScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MemorizeScaffold(
+      body: MLText("Recite this verse"),
+    );
+  }
+}
