+++ lib/widgets/compound_widgets/memorize_scaffold.dart
@@ -2,10 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:memory_lamp/theming/defaults.dart';
 import 'package:memory_lamp/widgets/ml_text.dart';
 
-class ReciteThisVerseScreen extends StatelessWidget {
+class MemorizeScaffold extends StatelessWidget {
   static const String routeName = "/reciteThisVerse";
 
-  const ReciteThisVerseScreen({Key? key}) : super(key: key);
+  final Widget body;
+
+  const MemorizeScaffold({required this.body, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -14,7 +16,7 @@ class ReciteThisVerseScreen extends StatelessWidget {
       body: SafeArea(
         child: Padding(
           padding: MLDefaults.screenPadding,
-          child: MLText("Recite this verse"),
+          child: body,
         ),
       ),
     );
